requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==43.0.1
datasets==3.0.0
dill==0.3.8
dnspython==2.6.1
fastapi==0.115.0
filelock==3.16.1
firebase-admin==6.5.0
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gcloud==0.18.3
gevent==24.2.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
jws==0.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numpy==2.1.1
oauth2client==4.1.3
packaging==24.1
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.1
pyparsing==3.1.4
Pyrebase==3.0.27
Pyrebase4==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tldextract==5.1.2
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.30.6
xxhash==3.5.0
yarl==1.12.1
zope.event==5.0
zope.interface==7.0.3